# public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "VibePass API",
    "description": "API endpoints for the VibePass Event Booking & OTT Streaming Explorer Platform.\n\n**Docs**: [/docs](/docs) (Swagger UI) · [/redoc](/redoc) (ReDoc) · [/openapi.json](/openapi.json) (schema)",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/payment/create-intent": {
      "post": {
        "summary": "Create Razorpay Order",
        "description": "Creates a Razorpay order on the server and returns the `orderId`, `amount`, `currency`, and `keyId` needed to initialise the Razorpay Checkout popup on the frontend.",
        "tags": [
          "Checkout & Payments"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "amount"
                ],
                "properties": {
                  "amount": {
                    "type": "number",
                    "description": "Decimal order amount in rupees (e.g. 500.5)",
                    "example": 500.5
                  },
                  "currency": {
                    "type": "string",
                    "default": "INR",
                    "example": "INR"
                  },
                  "bookingId": {
                    "type": "string",
                    "example": "45"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Razorpay order created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "orderId": {
                      "type": "string",
                      "example": "order_OFj1234abcd"
                    },
                    "amount": {
                      "type": "integer",
                      "description": "Amount in paise",
                      "example": 50000
                    },
                    "currency": {
                      "type": "string",
                      "example": "INR"
                    },
                    "keyId": {
                      "type": "string",
                      "description": "Razorpay publishable key for the checkout popup",
                      "example": "rzp_test_xxxxxx"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid amount",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Razorpay gateway error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/payment/verify": {
      "post": {
        "summary": "Verify Razorpay Payment",
        "description": "Verifies the HMAC-SHA256 signature of a completed Razorpay payment to confirm authenticity before booking confirmation. Called after the Razorpay checkout popup returns a successful `handler` callback.",
        "tags": [
          "Checkout & Payments"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "razorpay_order_id",
                  "razorpay_payment_id",
                  "razorpay_signature"
                ],
                "properties": {
                  "razorpay_order_id": {
                    "type": "string",
                    "example": "order_OFj1234abcd"
                  },
                  "razorpay_payment_id": {
                    "type": "string",
                    "example": "pay_OFj5678efgh"
                  },
                  "razorpay_signature": {
                    "type": "string",
                    "example": "9ef4dabb0070117e..."
                  },
                  "bookingId": {
                    "type": "string",
                    "example": "45"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Payment verified successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "paymentId": {
                      "type": "string",
                      "example": "pay_OFj5678efgh"
                    },
                    "orderId": {
                      "type": "string",
                      "example": "order_OFj1234abcd"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Signature mismatch or missing fields",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": false },
                    "error": { "type": "string", "example": "Payment verification failed. Invalid signature." }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error during verification",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "Error message explanation details."
          }
        }
      },
...
    }
  }
}